+++ packages/react/src/ui/value-display/types/text/index.tsx
@@ -1,1 +1,2 @@
 export * from "./text"
+export { TextEditor } from "./text-editor"